feat: integrate REV through bore encoder
Improve CSDutyCycleEncoder to handle real units for `setDistancePerRotation` and `setPositionOffset` instead of requiring values be between 0-1. 

`CSFusedEncoder` will now reset the relative encoder if it gets out of sync with the absolute encoder. This is mostly to handle inconsistent values during startup.

Tweaks to the ArmRotateMap to align the limits with the new readings, as well as adjust PID to work better with the new sensor. 

ArmRotate now uses the REV Through Bore encoder throughout instead of the encoder built into the motor.